requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
discord.py==1.2.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
multidict==4.5.2
praw==6.4.0
prawcore==1.0.1
pycodestyle==2.5.0
pylint==2.3.1
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
update-checker==0.16
urllib3==1.25.3
websocket-client==0.56.0
websockets==6.0
wrapt==1.11.2
yarl==1.3.0